+++ src/components/editor/markdown-renderer/markdown-renderer.tsx
@@ -36,10 +36,10 @@ import { replaceVimeoLink } from './regex-plugins/replace-vimeo-link'
 import { replaceYouTubeLink } from './regex-plugins/replace-youtube-link'
 import { ComponentReplacer, SubNodeConverter } from './replace-components/ComponentReplacer'
 import { GistReplacer } from './replace-components/gist/gist-replacer'
-import { HighlightedCodeReplacer } from './replace-components/highlighted-fence/highlighted-fence'
+import { HighlightedCodeReplacer } from './replace-components/highlighted-fence/highlighted-fence-replacer'
 import { MathjaxReplacer } from './replace-components/mathjax/mathjax-replacer'
 import { PdfReplacer } from './replace-components/pdf/pdf-replacer'
-import { QuoteOptionsReplacer } from './replace-components/quote-options/quote-options'
+import { QuoteOptionsReplacer } from './replace-components/quote-options/quote-options-replacer'
 import { TocReplacer } from './replace-components/toc/toc-replacer'
 import { VimeoReplacer } from './replace-components/vimeo/vimeo-replacer'
 import { YoutubeReplacer } from './replace-components/youtube/youtube-replacer'
