+++ src/components/editor/markdown-renderer/replace-components/highlighted-fence/highlighted-fence-replacer.tsx
